requirements.txt
torch==1.2.0
torchvision==0.4.0
numpy==1.18.0
Pillow==6.2.2
lmdb
easydict
pyfasttext
editdistance
tensorboardX
IPython
pyyaml
scipy
matplotlib
tqdm
opencv-python